Move Document out of pages
